Do not save 'birds.png' in Git.
`novice-r-practice.Rmd` shows how to save a file in R Markdown. The file (`birds.png`) was being regenerated every time the book was rebuilt, and Git thought it was different each time (presumably because of an embedded timestamp). This PR deletes the file and modifies `.gitignore` to ignore it in future.