Removing child document inclusion of questions.
The book uses child document inclusions for learning objects and keypoints so that we can automatically generate appendices that are consistent with the chapters' contents. Prior to this PR, we used child document inclusions for each chapter's motivating questions as well. However:

1.  These aren't collected in an appendix, and an inclusion that is only used once is probably better inline.
2.  If we write section headings in each chapter as questions, the inclusions would be redundant.

This PR therefore inlines the content of the `questions/*.md` files as a step toward renaming files to reflect novice/intermediate content. The intention is that these questions will eventually be removed entirely (redundant w/ section headings).