Append backup ext instead of replacing existing one
I believe this is the more common usage,
and it conveniently preserves information about the filetype.
This is also how the loop currently works,
and if it is to be rewritten to match the old text,
we need to include bash substring matching,
which I think is a bit too advanced.